Update batch_file_rename.py
The replace function replace all entries in the file name. While it is unlikely that a file has a name in the form of filename.ext.ext, it is theoretically possible. Maybe it would be better to change only the extension, because that's what the description promises. To do this we only need to store not only the extension of the file name, but the left part as well (line 20).